Refactor `sandbox` to share async/sync logic
Apply the iter-coroutine pattern (same as blob in a5bc8ca)
to eliminate ~600 lines of duplicated sync/async code.

All API operations now live as async methods on
BaseSandboxOpsClient, driven by SyncTransport (via
iter_coroutine) or AsyncTransport. Streaming log endpoints
remain as native sync/async generators since they cannot
use the iter-coroutine pattern.
